corrected a mistake in get_operator()

+++ helping_functions.py
@@ -29,9 +29,9 @@ def get_operator(hilbert, L, operator = None, symmetric = True):
                 name_fast = 'Ferro_correlation_function' + str(i)
                 observables[name_fast] = observ_fast
         else:
-            for i in range(1, L/2.):
+            for i in range(1, int(L/2.)):
                 observ_fast = operators.FerroCorrelationZ(hilbert=hilbert, j=L/2-i, k=L/2+i)
-                name_fast = 'Ferro_correlation_function' + str(2*i) #because k-j=2*i
+                name_fast = 'Ferro_correlation_function' + str(int(2*i)) #because k-j=2*i
                 observables[name_fast] = observ_fast
     elif(operator == 'StringCorr'):
         for i in range(1, L):
